// apps/web/src/types/generated/PeriodGenerationParameters.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AnnualTrendConfig } from "./AnnualTrendConfig";
import type { PeriodNoiseConfig } from "./PeriodNoiseConfig";
import type { PeriodVariableExpenseRule } from "./PeriodVariableExpenseRule";

export type PeriodGenerationParameters = {
  initial_revenue: string;
  initial_fixed_expense: string;
  revenue_trend: AnnualTrendConfig;
  fixed_expense_trend: AnnualTrendConfig;
  revenue_noise: PeriodNoiseConfig;
  fixed_expense_noise: PeriodNoiseConfig;
  demand_sensitivity_bp: number;
  variable_expense: PeriodVariableExpenseRule;
};