// apps/web/src/types/generated/AnnualTrendConfig.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type AnnualTrendConfig =
  | { "kind": "Fixed"; annual_growth_bp: number }
  | {
    "kind": "Persistent";
    annual_growth_min_bp: number;
    annual_growth_max_bp: number;
    duration_min_months: number;
    duration_max_months: number;
  };